# offchain/src/routes/shareRoutes.ts
// offchain/src/routes/shareRoutes.ts
// REST API routes for share token operations

import express from 'express';
import { getTokenLauncher } from '../shares';
import { getPassportManager } from '../services/passport/passportManager';
import { getPermanentStorage } from '../storage/depin';

export const shareRouter = express.Router();

/**
 * POST /v1/passports/:id/token/launch
 * Launch a share token for a passport
 */
shareRouter.post('/v1/passports/:id/token/launch', async (req, res) => {
  try {
    const passportId = req.params.id;
    const manager = getPassportManager();
    const passport = await manager.getPassport(passportId);

    if (!passport.ok || !passport.data) {
      return res.status(404).json({ error: 'Passport not found' });
    }

    const { name, symbol, totalSupply, decimals } = req.body;
    if (!name || !symbol || !totalSupply) {
      return res.status(400).json({ error: 'name, symbol, and totalSupply are required' });
    }

    // Upload token metadata to DePIN
    const metadataJson = {
      name,
      symbol,
      description: `Share token for Lucid ${passport.data.type} passport: ${passport.data.name || passportId}`,
      passport_id: passportId,
      passport_type: passport.data.type,
    };

    const uploadResult = await getPermanentStorage().uploadJSON(metadataJson, {
      tags: { 'Content-Type': 'application/json', 'lucid-share-token': 'true' },
    });

    const launcher = getTokenLauncher();
    const result = await launcher.launchToken({
      passportId,
      name,
      symbol,
      uri: uploadResult.url,
      totalSupply: Number(totalSupply),
      decimals: decimals ? Number(decimals) : undefined,
      owner: passport.data.owner,
    });

    // Store token mint on passport
    await manager.updatePassport(passportId, {
      metadata: { ...passport.data.metadata, share_token_mint: result.mint },
    });

    res.status(201).json(result);
  } catch (error) {
    res.status(500).json({ error: error instanceof Error ? error.message : 'Token launch failed' });
  }
});

/**
 * GET /v1/passports/:id/token
 * Get token info for a passport
 */
shareRouter.get('/v1/passports/:id/token', async (req, res) => {
  try {
    const passportId = req.params.id;
    const launcher = getTokenLauncher();
    const info = await launcher.getTokenInfo(passportId);

    if (!info) {
      return res.status(404).json({ error: 'No share token found for this passport' });
    }

    res.json(info);
  } catch (error) {
    res.status(500).json({ error: error instanceof Error ? error.message : 'Failed to get token info' });
  }
});

/**
 * POST /v1/passports/:id/token/airdrop
 * Trigger revenue airdrop for share token holders (admin only)
 */
shareRouter.post('/v1/passports/:id/token/airdrop', async (req, res) => {
  try {
    const passportId = req.params.id;
    const { amountLamports } = req.body;

    if (!amountLamports || Number(amountLamports) <= 0) {
      return res.status(400).json({ error: 'amountLamports must be a positive number' });
    }

    // Get the token mint from the passport
    const manager = getPassportManager();
    const passport = await manager.getPassport(passportId);
    if (!passport.ok || !passport.data) {
      return res.status(404).json({ error: 'Passport not found' });
    }

    const tokenMint = passport.data.share_token_mint || passport.data.metadata?.share_token_mint;
    if (!tokenMint) {
      return res.status(400).json({ error: 'No share token found for this passport. Launch a token first.' });
    }

    const { runRevenueAirdrop } = require('../jobs/revenueAirdrop');
    const result = await runRevenueAirdrop(passportId, tokenMint, Number(amountLamports));

    res.json(result);
  } catch (error) {
    res.status(500).json({ error: error instanceof Error ? error.message : 'Airdrop failed' });
  }
});