# openapi.yaml
openapi: 3.0.3
info:
  title: LucidLayer API
  version: 0.2.0
  description: |
    LucidLayer Offchain API.

    This OpenAPI spec is the source of truth for the actual backend routes.
    It is used to generate SDK clients via Speakeasy.

    Route groups:
    - `/v1/*` — Passports, Match, Run, Receipts, Epochs, Payouts, Compute
    - `/health/*` — System health and dependency checks
    - `/api/agents/*` — AI Agent MMR, Planner, and Orchestrator
servers:
  - url: https://api.lucid.foundation
    description: Production server

# Speakeasy MCP tool generation configuration
...
paths:
...
  /v1/passports/{passport_id}/token/launch:
    post:
      tags: [Shares]
      operationId: lucid_launch_share_token
      summary: Launch a share token for a passport
      x-speakeasy-mcp:
        name: lucid_launch_share_token
        description: Create and launch a share token (SPL) for fractional ownership of a passport.
      parameters:
        - in: path
          name: passport_id
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TokenLaunchRequest'
      responses:
        '201':
          description: Token launched
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenLaunchResponse'
        '400': { $ref: '#/components/responses/BadRequest' }
        '404': { $ref: '#/components/responses/NotFound' }
        '500': { $ref: '#/components/responses/InternalServerError' }
  /v1/passports/{passport_id}/token:
    get:
      tags: [Shares]
      operationId: lucid_get_share_token
      summary: Get share token info for a passport
      x-speakeasy-mcp:
        name: lucid_get_share_token
        description: Get token info (mint, supply, holders) for a passport's share token.
      parameters:
        - in: path
          name: passport_id
          required: true
          schema: { type: string }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenInfo'
        '404': { $ref: '#/components/responses/NotFound' }
        '500': { $ref: '#/components/responses/InternalServerError' }
  /v1/passports/{passport_id}/token/airdrop:
    post:
      tags: [Shares]
      operationId: lucid_trigger_revenue_airdrop
      summary: Trigger revenue airdrop for share token holders
      x-speakeasy-mcp:
        name: lucid_trigger_revenue_airdrop
        description: Snapshot token holders and distribute SOL proportionally based on holdings.
      parameters:
        - in: path
          name: passport_id
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AirdropRequest'
      responses:
        '200':
          description: Airdrop executed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AirdropResponse'
        '400': { $ref: '#/components/responses/BadRequest' }
        '404': { $ref: '#/components/responses/NotFound' }
        '500': { $ref: '#/components/responses/InternalServerError' }
...
  /v1/receipts:
    post:
      tags: [Receipts]
      operationId: lucid_create_receipt
      summary: Create a receipt
      x-speakeasy-mcp:
        name: lucid_create_receipt
        description: Create a receipt and append it to the receipt MMR.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateReceiptRequest'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                required: [success, receipt]
                properties:
                  success: { type: boolean }
                  receipt: { $ref: '#/components/schemas/Receipt' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '500': { $ref: '#/components/responses/InternalServerError' }
  /v1/receipts/{receipt_id}:
    get:
      tags: [Receipts]
      operationId: lucid_get_receipt
      summary: Get a receipt
      x-speakeasy-mcp:
        name: lucid_get_receipt
        description: Retrieve a receipt by receipt_id.
      parameters:
        - in: path
          name: receipt_id
          required: true
          schema: { type: string }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                required: [success, receipt]
                properties:
                  success: { type: boolean }
                  receipt: { $ref: '#/components/schemas/Receipt' }
        '404': { $ref: '#/components/responses/NotFound' }
        '500': { $ref: '#/components/responses/InternalServerError' }
  /v1/receipts/{receipt_id}/verify:
    get:
      tags: [Receipts]
      operationId: lucid_verify_receipt
      summary: Verify a receipt (hash + signature + inclusion)
      x-speakeasy-mcp:
        name: lucid_verify_receipt
        description: Verify a receipt (hash + signature + inclusion proof).
      parameters:
        - in: path
          name: receipt_id
          required: true
          schema: { type: string }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReceiptVerification'
        '404': { $ref: '#/components/responses/NotFound' }
        '500': { $ref: '#/components/responses/InternalServerError' }
  /v1/receipts/{receipt_id}/proof:
    get:
      tags: [Receipts]
      operationId: lucid_get_receipt_proof
      summary: Get inclusion proof for receipt
      x-speakeasy-mcp:
        name: lucid_get_receipt_proof
        description: Get the inclusion proof for a receipt.
      parameters:
        - in: path
          name: receipt_id
          required: true
          schema: { type: string }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                required: [success, proof]
                properties:
                  success: { type: boolean }
                  proof: { $ref: '#/components/schemas/ReceiptProof' }
        '404': { $ref: '#/components/responses/NotFound' }
        '500': { $ref: '#/components/responses/InternalServerError' }
  /v1/verify/{receipt_hash}:
    get:
      tags: [Receipts]
      operationId: lucid_verify_receipt_by_hash
      summary: Verify receipt by hash with inclusion proof and epoch info
      x-speakeasy-mcp:
        name: lucid_verify_receipt_by_hash
        description: >
          Verify a receipt by its hash. Returns inclusion proof and on-chain
          anchoring status. Critical P0.9 endpoint for Fluid Compute.
      parameters:
        - in: path
          name: receipt_hash
          required: true
          schema: { type: string, minLength: 64, maxLength: 64 }
          description: 64 hex character receipt hash
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReceiptHashVerification'
        '400': { $ref: '#/components/responses/BadRequest' }
        '404': { $ref: '#/components/responses/NotFound' }
        '500': { $ref: '#/components/responses/InternalServerError' }
  /v1/mmr/root:
    get:
      tags: [Receipts]
      operationId: lucid_get_mmr_root
      summary: Get current MMR root
      x-speakeasy-mcp:
        name: lucid_get_mmr_root
        description: Get the current MMR root and leaf count.
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                required: [success, root, leaf_count]
                properties:
                  success: { type: boolean }
                  root: { type: string }
                  leaf_count: { type: integer }
        '500': { $ref: '#/components/responses/InternalServerError' }
  /v1/signer/pubkey:
    get:
      tags: [Receipts]
      operationId: lucid_get_signer_pubkey
      summary: Get orchestrator signing public key
      x-speakeasy-mcp:
        name: lucid_get_signer_pubkey
        description: Get the orchestrator signing public key.
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                required: [success, pubkey]
                properties:
                  success: { type: boolean }
                  signer_type: { type: string }
                  pubkey: { type: string }
        '500': { $ref: '#/components/responses/InternalServerError' }

  # ===========================================================================
  # Epochs / anchoring
  # ===========================================================================
...
components:
  responses:
    BadRequest:
      description: Bad Request
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
...
    NotFound:
      description: Not Found
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
...
    InternalServerError:
      description: Internal Server Error
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        success: { type: boolean }
        error: { type: string }
        message: { type: string }
        error_code: { type: string }
        details: {}

    # =========================================================================
    # Compute
    # =========================================================================
...
    Receipt:
      type: object
      required: [run_id, model_passport_id, compute_passport_id, policy_hash, runtime, timestamp, receipt_hash, receipt_signature, signer_pubkey, signer_type, metrics]
      properties:
        schema_version: { type: string }
        run_id: { type: string }
        model_passport_id: { type: string }
        compute_passport_id: { type: string }
        policy_hash: { type: string }
        runtime: { type: string }
        timestamp: { type: integer }
        trace_id: { type: string }
        image_hash: { type: string }
        model_hash: { type: string }
        attestation: { type: object, additionalProperties: true }
        execution_mode: { type: string }
        job_hash: { type: string }
        quote_hash: { type: string }
        node_id: { type: string }
        runtime_hash: { type: string, nullable: true }
        gpu_fingerprint: { type: string, nullable: true }
        outputs_hash: { type: string }
        output_ref: { type: string }
        start_ts: { type: integer }
        end_ts: { type: integer }
        input_ref: { type: string }
        error_code: { type: string }
        error_message: { type: string }
        metrics:
          type: object
          required: [ttft_ms, tokens_in, tokens_out]
          properties:
            ttft_ms: { type: integer }
            tokens_in: { type: integer }
            tokens_out: { type: integer }
            p95_ms: { type: integer }
            total_latency_ms: { type: integer }
            queue_wait_ms: { type: integer }
            queue_time_ms: { type: integer }
            cold_start_ms: { type: integer }
            model_load_ms: { type: integer }
            cache_hit: { type: boolean }
        receipt_hash: { type: string }
        receipt_signature: { type: string }
        signer_pubkey: { type: string }
        signer_type:
          type: string
          enum: [orchestrator, compute, worker]
        _mmr_leaf_index: { type: integer }
        anchor:
          type: object
          nullable: true
          properties:
            chain: { type: string }
            tx: { type: string }
            root: { type: string }
            epoch_id: { type: string }
    CreateReceiptRequest:
      type: object
      description: Input fields for creating a receipt. The backend computes the hash, signature, and wraps metrics.
      required: [model_passport_id, compute_passport_id, policy_hash, runtime, tokens_in, tokens_out, ttft_ms]
      properties:
        model_passport_id: { type: string }
        compute_passport_id: { type: string }
        policy_hash: { type: string }
        runtime: { type: string }
        tokens_in: { type: integer }
        tokens_out: { type: integer }
        ttft_ms: { type: integer }
        total_latency_ms: { type: integer }
        timestamp: { type: integer }
        trace_id: { type: string }
        run_id: { type: string }
        image_hash: { type: string }
        model_hash: { type: string }
        attestation: { type: object, additionalProperties: true }
        execution_mode: { type: string }
        node_id: { type: string }
        runtime_hash: { type: string }
        gpu_fingerprint: { type: string }
    ReceiptProof:
      type: object
      required: [run_id, receipt_hash, leaf_index, proof, root]
      properties:
        run_id: { type: string }
        receipt_hash: { type: string }
        leaf_index: { type: integer }
        proof:
          type: array
          items: { type: string }
          description: Sibling hashes along the Merkle path
        root: { type: string }
        directions:
          type: array
          items:
            type: string
            enum: ['L', 'R']
          description: Direction of each sibling (L = sibling is on left, R = on right).
    ReceiptVerification:
      type: object
      properties:
        success: { type: boolean }
        valid: { type: boolean }
        hash_valid: { type: boolean }
        signature_valid: { type: boolean }
        inclusion_valid: { type: boolean }
        expected_hash: { type: string }
        computed_hash: { type: string }
        merkle_root: { type: string }
    ReceiptHashVerification:
      type: object
      required: [success, verified, receipt_hash]
      properties:
        success: { type: boolean }
        verified: { type: boolean }
        receipt_hash: { type: string }
        run_id: { type: string }
        hash_valid: { type: boolean }
        signature_valid: { type: boolean }
        signer_pubkey: { type: string }
        signer_type: { type: string }
        execution_mode: { type: string }
        runtime_hash: { type: string }
        gpu_fingerprint: { type: string }
        inclusion_proof:
          type: object
          nullable: true
          properties:
            leaf_index: { type: integer }
            proof:
              type: array
              items: { type: string }
            root: { type: string }
            directions:
              type: array
              items: { type: string }
        inclusion_valid: { type: boolean }
        epoch:
          type: object
          nullable: true
          properties:
            epoch_id: { type: string }
            mmr_root: { type: string }
            chain_tx: { type: string }
            finalized_at: { type: integer }
        on_chain_verified: { type: boolean }
        tx_signature: { type: string, nullable: true }

    # =========================================================================
    # Epoch schemas
    # =========================================================================
...
    TokenLaunchRequest:
      type: object
      required: [name, symbol, totalSupply]
      properties:
        name: { type: string, description: 'Token name (e.g., "Llama 3.1 Share")' }
        symbol: { type: string, description: 'Token symbol (e.g., "LLAMA")' }
        totalSupply: { type: integer, minimum: 1, description: 'Total token supply' }
        decimals: { type: integer, minimum: 0, maximum: 18, default: 6, description: 'Token decimals' }
    TokenLaunchResponse:
      type: object
      required: [mint, txSignature, totalSupply, provider]
      properties:
        mint: { type: string, description: 'SPL token mint address' }
        txSignature: { type: string, description: 'Transaction signature' }
        totalSupply: { type: integer }
        provider: { type: string, description: 'Token launcher provider used' }
    TokenInfo:
      type: object
      required: [mint, passportId, name, symbol, totalSupply, decimals]
      properties:
        mint: { type: string, description: 'SPL token mint address' }
        passportId: { type: string }
        name: { type: string }
        symbol: { type: string }
        totalSupply: { type: integer }
        decimals: { type: integer }
        holders: { type: integer, description: 'Number of unique token holders' }
    AirdropRequest:
      type: object
      required: [amountLamports]
      properties:
        amountLamports: { type: integer, minimum: 1, description: 'SOL amount in lamports to distribute' }
    AirdropResponse:
      type: object
      required: [passportId, tokenMint, totalDistributed, holders, distributions, txSignatures]
      properties:
        passportId: { type: string }
        tokenMint: { type: string }
        totalDistributed: { type: integer, description: 'Total lamports distributed' }
        holders: { type: integer, description: 'Number of holders who received airdrop' }
        distributions:
          type: array
          items:
            type: object
            properties:
              holder: { type: string }
              balance: { type: number }
              share: { type: number }
              amountLamports: { type: integer }
        txSignatures:
          type: array
          items: { type: string }